refactor : session 의 null 확인 위치 변경
